refactor: Calculator 제거 및 메서드 board와 chessgame으로 분배
