+++ src/main/java/softeer2nd/Main.java
@@ -1,5 +1,7 @@
 package softeer2nd;
 
+import softeer2nd.chess.simulator.ChessSimulator;
+
 public class Main {
     public static void main(String[] args) {
