+++ src/main/java/softeer2nd/chess/game/ChessGame.java
@@ -37,6 +37,39 @@ public class ChessGame {
         return posMap;
     }
 
+    public double calculatePoint(Piece.Color color) {
+        return board.stream()
+                .mapToDouble(rank -> rank.calculatePoint(color))
+                .sum() + calculatePawn(color);
+    }
+
+    private double calculatePawn(Piece.Color color) {
+
+        double result = 0;
+
+        for (int i = 0; i < 8; i++) {
+            int pawnCount = 0;
+
+            for (int j = 0; j < 8; j++) {
+                Piece piece = board.get(j).get(i);
+
+                if (piece.getType().equals(Piece.Type.PAWN) && piece.getColor().equals(color)) {
+                    pawnCount++;
+                }
+            }
+
+            if (pawnCount == 1) {
+                result += 1;
+            }
+
+            if (pawnCount > 1) {
+                result += (pawnCount * 0.5);
+            }
+        }
+
+        return result;
+    }
+
     public void move(String sourcePosition, String targetPosition) {
 
         Map<String, Integer> sourcePosMap = getPosition(sourcePosition);
